mcp_google_server-z/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-api-python-client
google-auth
python-dotenv